Pull out KeyFactory
