+++ .idea/compiler.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CompilerConfiguration">
-    <bytecodeTargetLevel target="1.8">
+    <bytecodeTargetLevel target="11">
       <module name="common_jvmMain" target="1.8" />
       <module name="common_jvmTest" target="1.8" />
       <module name="data-donation-sdk.buildSrc" target="1.8" />
