+++ data-donation-crypto/src/commonMain/kotlin/care/data4life/datadonation/crypto/KeyFactory.kt
@@ -16,19 +16,8 @@
 
 package care.data4life.datadonation.crypto
 
-actual class CryptoService actual constructor() : CryptoContract.Service {
-    actual override fun encrypt(
-        payload: ByteArray,
-        publicKey: String
-    ): ByteArray {
-        TODO()
-    }
+import care.data4life.datadonation.crypto.model.KeyPair
 
-    actual override fun sign(
-        payload: ByteArray,
-        privateKey: String,
-        saltLength: Int,
-    ): ByteArray {
-        TODO()
-    }
+expect object KeyFactory : CryptoContract.KeyFactory {
+    override fun createKeyPair(): KeyPair
 }
